added patient history screen

+++ arduino_code/temp_app_arduino/functions.ino
@@ -8,7 +8,7 @@ void send_temp(){
    SendToProcessing('T',avg);
   
   }
-void SendToProcessing(char a,int data){
+void SendToProcessing(char a,float data){
   Serial.print(a);
   Serial.println(data);
